requirements.txt
aiohttp==3.9.5            # via discord-py
aiosignal==1.3.1          # via aiohttp
attrs==23.2.0             # via aiohttp
discord-py==2.4.0         # via -r requirements.in
frozenlist==1.4.1         # via aiohttp, aiosignal
idna==3.7                 # via yarl
multidict==6.0.5          # via aiohttp, yarl
python-dotenv==1.0.1      # via -r requirements.in
yarl==1.9.4               # via aiohttp
